requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
supabase==2.0.2
pydantic==2.5.0
PyMuPDF==1.23.8
docx2txt==0.8